+++ src/main/java/com/example/domain/repository/MissionRepository.java
@@ -4,7 +4,7 @@ import com.example.domain.model.MoonMission;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface MissionRepository extends JpaRepository<MoonMission,Short> {
+public interface MissionRepository extends JpaRepository<MoonMission,Integer> {
     @Query("select count(m) from MoonMission m where extract(year from m.launchDate) = :year")
     long countMissionsByYear(int year);
 }
